# Extract inline vars from the status handler